issue 10: fixed mypy type hints and reformatted
